Suppress Material theme's font support
For our project I think it's better not to rely on Google fonts
(although they look nice ...), reducing the hassle of working
towards GDPR compliance.

It should be thought about another way to customize the
look and feel by providing a web font that is directly
server by the server.